+++ js/src/test/scala/com/raquo/domtypes/CompileTest.scala
@@ -8,6 +8,7 @@ import com.raquo.domtypes.generic.defs.props.Props
 import com.raquo.domtypes.generic.defs.reflectedAttrs.ReflectedAttrs
 import com.raquo.domtypes.generic.defs.styles.{Styles, Styles2}
 import com.raquo.domtypes.generic.keys.{Attr, EventProp, Prop, Style}
+import com.raquo.domtypes.jsdom.defs.InputElementTargetEvent
 import com.raquo.domtypes.jsdom.defs.eventProps.{ClipboardEventProps, ErrorEventProps, FormEventProps, KeyboardEventProps, MediaEventProps, MiscellaneousEventProps, MouseEventProps, WindowEventProps}
 import com.raquo.domtypes.jsdom.defs.tags.{DocumentTags, EmbedTags, FormTags, GroupingTags, MiscTags, SectionTags, TableTags, TextTags}
 import org.scalajs.dom
@@ -38,7 +39,7 @@ class CompileTest {
     // Event Props
     with ClipboardEventProps[EventProp]
     with ErrorEventProps[EventProp]
-    with FormEventProps[EventProp]
+    with FormEventProps[EventProp, InputElementTargetEvent] // Note: InputElementTargetEvent is not a 100% type safe choice. See README.
     with KeyboardEventProps[EventProp]
     with MediaEventProps[EventProp]
     with MiscellaneousEventProps[EventProp]